requirements.txt
Flask==2.2.3
pexelsPy==1.0.1
python-dotenv==1.0.0
requests==2.27.1
tqdm==4.64.1
